refactor(gateway): extract voice/STT/vision/media methods to a mixin (mechanical move)
gateway/run.py is 26,877 lines and GatewayRunner alone is 20,131 of them.
This continues the decomposition that already produced
GatewayAuthorizationMixin, GatewayKanbanWatchersMixin and
GatewaySlashCommandsMixin, moving one coherent concern out to
gateway/media.py: everything that turns inbound audio and images into text
the agent can read, and turns a response back into voice or media on the
way out.

21 methods, ~970 lines — Discord voice-channel membership
(_handle_voice_channel_join/leave/input, _handle_voice_timeout_cleanup,
_voice_channel_sidecar_note), the /voice mode state (_voice_key,
_load_voice_modes, _save_voice_modes, _sync_voice_mode_state_to_adapter),
the STT pipeline and its echo-once bookkeeping (_pending_event_audio_paths,
_transcribe_pending_audio_event_once, _echo_pending_stt_transcripts_once,
_transcribe_and_echo_pending_voice, _is_duplicate_voice_transcript,
_should_echo_stt_transcripts), inbound enrichment
(_enrich_message_with_transcription, _enrich_message_with_vision), and
outbound media/voice (_should_send_voice_reply, _send_voice_reply,
_deliver_media_from_response, _warn_if_docker_media_delivery_is_risky).

Behaviour-neutral by construction: all 21 method bodies are byte-identical
to their previous text, verified by an AST comparison that strips exactly
the injected import statements and nothing else. The only edits are three
deferred `from gateway.run import ...` lines for the four module-level
names the bodies still reach for (_DOCKER_VOLUME_SPEC_RE,
_DOCKER_MEDIA_OUTPUT_CONTAINER_PATHS, _event_media_is_stt_input,
_probe_audio_duration) — resolved at call time, the pattern the earlier
phases established, so this module never imports gateway.run, nothing
circularises, and a test patching gateway.run.X still takes effect. The
media-classifier family stays in run.py intact rather than being split
across two files. The module binds logger = logging.getLogger("gateway.run")
so extracted log records keep their original logger name, matching
gateway/kanban_watchers.py, and takes a module-private _UNSET sentinel the
way gateway/status.py already does, since that name is a keyword default
compared only against itself.

run.py: 26,877 -> 25,913 lines.

Rebuilt on current main rather than rebased: the #75619 branch was ~2,000
commits behind and gateway/run.py had drifted +1,192/-81 lines. For a
mechanical move, re-deriving the extraction guarantees the moved bodies
match today's main byte for byte, which a hand-resolved conflict cannot.

test_load_voice_modes_skips_legacy_keys asserted the warning by patching
the gateway.run.logger module global, which the move relocates. Switched to
caplog on logger "gateway.run", which is what the code actually emits — it
asserts the behaviour instead of which module object holds the logger, and
it passes unchanged on both this branch and pristine main.
